Makes 1 out of 2 matrices for LU. See update notes.
The new function prepareLU() in LU.cpp creates the matrix A in equation Ax = B where x is the matrix of unknown currents. See solution.pdf for a walkthrough of solving for the currents by hand vs. prepareLU() for sample circuit input2.txt.

Need to make matrix B to proceed with LU, but B is not a square matrix.